+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 All notable changes to the extension will be documented in this file.
 
-## [1.4.0] - 2020-12-23
+## [1.4.0] - 
 
 - Added @extends('base') tag for inheritance.
-- Added clearCache command.
-- Added more tests.
+- Added [CSS: Validate Attributes] command.
+- Added [CSS: Clear Cache] command.
+- Added integration tests.
 
 ## [1.3.3] - 2020-12-22